refator: CalendarEventBuilder 위치 변경
